Fix broken GitHub actions due to line lengths and pylint command typo.
