+++ .github/workflows/tests.yml
@@ -33,6 +33,11 @@ jobs:
         run: poetry install
         if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
 
-      - name: Run tests
+      - name: Activate venv
         run: |
-          poetry run python tests.py DataTest
+          echo "VIRTUAL_ENV=$GITHUB_WORKSPACE/.venv" >> $GITHUB_ENV
+          echo "PYTHONPATH=$GITHUB_WORKSPACE" >> $GITHUB_ENV
+          echo "$GITHUB_WORKSPACE/.venv/bin" >> $GITHUB_PATH
+
+      - name: Run tests
+        run: python -m unittest discover -s tests -p 'test_*.py'
