+++ pyproject.toml
@@ -28,7 +28,7 @@ version = "0.0.1"
 "Organisation Homepage" = "https://free.law/"
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.5"
 
 [tool.poetry.dev-dependencies]
 black = "^20.8b1"
